# api/apifrostline.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "apifrostline.proto",
    "description": "The realtime protocol for the template match engine.",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Nakama"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v2/getshopdata": {
      "get": {
        "operationId": "Nakama_GetShop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetShopData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Nakama"
        ]
      }
    },
    "/v2/gettransactioninfo": {
      "get": {
        "summary": "get spesific transactions for UserId",
        "operationId": "Nakama_GetTransactionInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetTransactionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Nakama"
        ]
      }
    },
    "/v2/getuserinfo": {
      "get": {
        "summary": "Get User Account details.",
        "operationId": "Nakama_GetUserInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Nakama"
        ]
      }
    },
    "/v2/getuserinventory": {
      "get": {
        "operationId": "Nakama_GetInventory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetUserInventory"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Nakama"
        ]
      }
    },
    "/v2/getusertransactions": {
      "get": {
        "summary": "Get purchased transactions",
        "operationId": "Nakama_GetTransactions",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetTransactionsList"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Nakama"
        ]
      }
    }
  },
  "definitions": {
    "apiGetShopData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiShopItem"
          }
        }
      }
    },
    "apiGetTransactionResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "apiGetTransactionsList": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "transactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiTransaction"
          }
        }
      }
    },
    "apiGetUserInventory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiInventoryItem"
          }
        }
      }
    },
    "apiInventoryItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        }
      }
    },
    "apiShopItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        }
      }
    },
    "apiTransaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "apiUserInfoResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "lang": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "timezone": {
          "type": "string"
        },
        "coins": {
          "type": "string"
        },
        "gems": {
          "type": "string"
        },
        "xp": {
          "type": "string"
        },
        "level": {
          "type": "string"
        },
        "lastLogin": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}